Add files via upload
I've added a json file and the file for tests. The only way I found that I can use the object in a different file it to create modules, though I'll keep researching because it's easier for me to write console tests in a separate file. Also I've added several functions  as per the assignment